refactor: remove circular route dependencies
