+++ src/pages/_navigation-meta.ts
@@ -1,0 +1,25 @@
+export const navigationItems = {
+  home: { label: 'layout.components.navigation.tabs.home', path: '/' },
+  proxies: {
+    label: 'layout.components.navigation.tabs.proxies',
+    path: '/proxies',
+  },
+  profiles: {
+    label: 'layout.components.navigation.tabs.profiles',
+    path: '/profile',
+  },
+  connections: {
+    label: 'layout.components.navigation.tabs.connections',
+    path: '/connections',
+  },
+  rules: { label: 'layout.components.navigation.tabs.rules', path: '/rules' },
+  logs: { label: 'layout.components.navigation.tabs.logs', path: '/logs' },
+  unlock: {
+    label: 'layout.components.navigation.tabs.unlock',
+    path: '/unlock',
+  },
+  settings: {
+    label: 'layout.components.navigation.tabs.settings',
+    path: '/settings',
+  },
+} as const
